+++ internal/pkg/scaffold/gopkgtoml_test.go
@@ -48,27 +48,27 @@ required = [
 
 [[override]]
   name = "k8s.io/api"
-  # revision for tag "kubernetes-1.13.4"
-  revision = "5cb15d34447165a97c76ed5a60e4e99c8a01ecfe"
+  # revision for tag "kubernetes-1.14.1"
+  revision = "6e4e0e4f393bf5e8bbff570acd13217aa5a770cd"
 
 [[override]]
   name = "k8s.io/apiextensions-apiserver"
-  # revision for tag "kubernetes-1.13.4"
-  revision = "d002e88f6236312f0289d9d1deab106751718ff0"
+  # revision for tag "kubernetes-1.14.1"
+  revision = "727a075fdec8319bf095330e344b3ccc668abc73"
 
 [[override]]
   name = "k8s.io/apimachinery"
-  # revision for tag "kubernetes-1.13.4"
-  revision = "86fb29eff6288413d76bd8506874fddd9fccdff0"
+  # revision for tag "kubernetes-1.14.1"
+  revision = "6a84e37a896db9780c75367af8d2ed2bb944022e"
 
 [[override]]
   name = "k8s.io/client-go"
-  # revision for tag "kubernetes-1.13.4"
-  revision = "b40b2a5939e43f7ffe0028ad67586b7ce50bb675"
+  # revision for tag "kubernetes-1.14.1"
+  revision = "1a26190bd76a9017e289958b9fba936430aa3704"
 
 [[override]]
   name = "github.com/coreos/prometheus-operator"
-  version = "=v0.29.0"
+  version = "=v0.31.1"
 
 [[override]]
   name = "k8s.io/kube-state-metrics"
@@ -76,7 +76,12 @@ required = [
 
 [[override]]
   name = "sigs.k8s.io/controller-runtime"
-  version = "=v0.1.12"
+  version = "=v0.2.0-beta.3"
+
+# Required when resolving controller-runtime dependencies.
+[[override]]
+  name = "gopkg.in/fsnotify.v1"
+  source = "https://github.com/fsnotify/fsnotify.git"
 
 [[constraint]]
   name = "github.com/operator-framework/operator-sdk"
@@ -91,5 +96,5 @@ required = [
   [[prune.project]]
     name = "k8s.io/kube-state-metrics"
     unused-packages = true
-  
+
 `