+++ internal/pkg/scaffold/helm/gopkgtoml.go
@@ -43,46 +43,46 @@ const gopkgTomlTmpl = `[[constraint]]
 
 [[override]]
   name = "k8s.io/api"
-  version = "kubernetes-1.13.4"
+  version = "kubernetes-1.14.1"
 
 [[override]]
   name = "k8s.io/apimachinery"
-  version = "kubernetes-1.13.4"
+  version = "kubernetes-1.14.1"
 
 [[override]]
   name = "k8s.io/apiextensions-apiserver"
-  version = "kubernetes-1.13.4"
+  version = "kubernetes-1.14.1"
 
 [[override]]
   name = "k8s.io/apiserver"
-  version = "kubernetes-1.13.4"
+  version = "kubernetes-1.14.1"
 
 [[override]]
   name = "k8s.io/client-go"
-  version = "kubernetes-1.13.4"
+  version = "kubernetes-1.14.1"
 
 [[override]]
   name = "k8s.io/cli-runtime"
-  version = "kubernetes-1.13.4"
+  version = "kubernetes-1.14.1"
 
 # We need overrides for the following imports because dep can't resolve them
 # correctly. The easiest way to get this right is to use the versions that
-# k8s.io/helm uses. See https://github.com/helm/helm/blob/v2.13.1/glide.lock
+# k8s.io/helm uses. See https://github.com/helm/helm/blob/v2.14.1/glide.lock
 [[override]]
   name = "k8s.io/kubernetes"
-  revision = "c6d339953bd4fd8c021a6b5fb46d7952b30be9f9"
+  revision = "b7394102d6ef778017f2ca4046abbaa23b88c290"
 
 [[override]]
-name = "github.com/russross/blackfriday"
-revision = "300106c228d52c8941d4b3de6054a6062a86dda3"
+  name = "github.com/russross/blackfriday"
+  revision = "300106c228d52c8941d4b3de6054a6062a86dda3"
 
 [[override]]
-name = "github.com/docker/distribution"
-revision = "edc3ab29cdff8694dd6feb85cfeb4b5f1b38ed9c"
+  name = "github.com/docker/distribution"
+  revision = "edc3ab29cdff8694dd6feb85cfeb4b5f1b38ed9c"
 
 [[override]]
-name = "github.com/docker/docker"
-revision = "a9fbbdc8dd8794b20af358382ab780559bca589d"
+  name = "github.com/docker/docker"
+  revision = "a9fbbdc8dd8794b20af358382ab780559bca589d"
 
 [prune]
   go-tests = true