+++ internal/pkg/scaffold/olm-catalog/csv_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/operator-framework/operator-sdk/internal/util/diffutil"
 	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
 
-	"github.com/coreos/go-semver/semver"
+	"github.com/blang/semver"
 	"github.com/ghodss/yaml"
 	olmapiv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
 	olminstall "github.com/operator-framework/operator-lifecycle-manager/pkg/controller/install"
@@ -136,9 +136,12 @@ func TestUpdateVersion(t *testing.T) {
 		t.Fatalf("Failed to update csv with version %s: (%v)", newCSVVer, err)
 	}
 
-	wantedSemver := semver.New(newCSVVer)
-	if !csv.Spec.Version.Equal(*wantedSemver) {
-		t.Errorf("Wanted csv version %v, got %v", *wantedSemver, csv.Spec.Version)
+	wantedSemver, err := semver.Parse(newCSVVer)
+	if err != nil {
+		t.Errorf("Failed to parse %s: %v", newCSVVer, err)
+	}
+	if !csv.Spec.Version.Equals(wantedSemver) {
+		t.Errorf("Wanted csv version %v, got %v", wantedSemver, csv.Spec.Version)
 	}
 	wantedName := getCSVName(operatorName, newCSVVer)
 	if csv.ObjectMeta.Name != wantedName {