+++ pkg/test/client.go
@@ -32,7 +32,7 @@ var _ FrameworkClient = &frameworkClient{}
 
 type FrameworkClient interface {
 	Get(gCtx goctx.Context, key dynclient.ObjectKey, obj runtime.Object) error
-	List(gCtx goctx.Context, opts *dynclient.ListOptions, list runtime.Object) error
+	List(gCtx goctx.Context, list runtime.Object, opts ...dynclient.ListOptionFunc) error
 	Create(gCtx goctx.Context, obj runtime.Object, cleanupOptions *CleanupOptions) error
 	Delete(gCtx goctx.Context, obj runtime.Object, opts ...dynclient.DeleteOptionFunc) error
 	Update(gCtx goctx.Context, obj runtime.Object) error
@@ -93,8 +93,8 @@ func (f *frameworkClient) Get(gCtx goctx.Context, key dynclient.ObjectKey, obj r
 	return f.Client.Get(gCtx, key, obj)
 }
 
-func (f *frameworkClient) List(gCtx goctx.Context, opts *dynclient.ListOptions, list runtime.Object) error {
-	return f.Client.List(gCtx, opts, list)
+func (f *frameworkClient) List(gCtx goctx.Context, list runtime.Object, opts ...dynclient.ListOptionFunc) error {
+	return f.Client.List(gCtx, list, opts...)
 }
 
 func (f *frameworkClient) Delete(gCtx goctx.Context, obj runtime.Object, opts ...dynclient.DeleteOptionFunc) error {